+++ src/internal_events/datadog_traces.rs
@@ -38,3 +38,27 @@ impl InternalEvent for DatadogTracesEncodingError {
         }
     }
 }
+
+#[derive(Debug)]
+pub struct DatadogTracesAPMStatsError<E> {
+    pub error: E,
+}
+
+impl<E: std::fmt::Display> InternalEvent for DatadogTracesAPMStatsError<E> {
+    fn emit(self) {
+        error!(
+            message = "Failed sending APM stats payload.",
+            error = %self.error,
+            error_type = error_type::WRITER_FAILED,
+            stage = error_stage::SENDING,
+            internal_log_rate_limit = true,
+        );
+        counter!(
+            "component_errors_total", 1,
+            "error_type" => error_type::WRITER_FAILED,
+            "stage" => error_stage::SENDING,
+        );
+
+        // No dropped events because APM stats payloads are not considered events.
+    }
+}
