+++ src/sinks/datadog/traces/mod.rs
@@ -11,11 +11,11 @@ mod integration_tests;
 #[cfg(test)]
 mod tests;
 
+pub(crate) mod apm_stats;
 mod config;
 mod request_builder;
 mod service;
 mod sink;
-mod stats;
 
 #[allow(warnings, clippy::pedantic, clippy::nursery)]
 pub(crate) mod ddsketch_full {
